+++ tests/models/dstability/test_dstability_model.py
@@ -43,42 +43,7 @@ from geolib.soils import (
 from tests.utils import TestUtils
 
 
-class TestDStabilityModel:
-    @staticmethod
-    def _get_standard_layers():
-        layer_1 = [
-            Point(x=-50, z=-10),
-            Point(x=50, z=-10),
-            Point(x=50, z=-20),
-            Point(x=-50, z=-20),
-        ]
-        layer_2 = [
-            Point(x=-50, z=-5),
-            Point(x=50, z=-5),
-            Point(x=50, z=-10),
-            Point(x=-50, z=-10),
-        ]
-        layer_3 = [
-            Point(x=-50, z=0),
-            Point(x=50, z=0),
-            Point(x=50, z=-5),
-            Point(x=-50, z=-5),
-        ]
-        embankment = [
-            Point(x=-10, z=0),
-            Point(x=0, z=2),
-            Point(x=10, z=2),
-            Point(x=30, z=0),
-        ]
-
-        layers_and_soils = [
-            (layer_1, "Sand"),
-            (layer_2, "H_Ro_z&k"),
-            (layer_3, "H_Rk_k_shallow"),
-            (embankment, "H_Aa_ht_old"),
-        ]
-        return layers_and_soils
-    
+class TestDStabilityModel: 
     @pytest.mark.unittest
     def test_instantiate_stability_model(self):
         assert isinstance(DStabilityModel(filename=None), BaseModel), (
@@ -408,7 +373,7 @@ class TestDStabilityModel:
     def test_generate_simple_model(self):
         dm = DStabilityModel()
 
-        layers_and_soils = self._get_standard_layers()
+        layers_and_soils = TestUtils._get_standard_layers()
 
         layer_ids = []
 
@@ -476,7 +441,7 @@ class TestDStabilityModel:
         soil_peat_id.mohr_coulomb_parameters.cohesion.mean = 0.5
         soil_peat_id = dm.add_soil(soil_peat_id)
 
-        layers_and_soils = self._get_standard_layers()
+        layers_and_soils = TestUtils._get_standard_layers()
 
         layer_ids = []
